refactor: split left and right
